requirements.txt
py-cord[voice]
urllib3
audioop-lts
GitPython